agent/requirements.txt
twilio
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic
elevenlabs>=1.0.0
python-dotenv>=1.0.0
pyngrok>=5.0.0
openai
weave